refactor: Update INotificationsService to accept options object
The INotificationsService interface has been updated to accept an options object for sending notifications. This allows for more flexibility in specifying the token, message, and data for each notification.

feat: Update FleetsManager to use options object for sending notifications

The FleetsManager class has been updated to use the new options object when sending notifications. This change improves the readability and maintainability of the code.

feat: Add data payload to notification messages

Notification messages now include a data payload that provides additional information about the notification type and related entities. This enhancement allows for more customized handling of notifications on the client side.